requirements.txt
fastapi
uvicorn
requests==2.31.0
Pillow==10.0.0
python-dotenv==1.0.0
python-multipart
google-auth
PyJWT
jwt
pyjwt